This is for a side, personal project
THIS IS A PERSONAL PROJECT NOT PROFESSIONAL (yet)

I want to track my moods, health, etc, and I didn't see any app that was as customizable as I hoped. So I made my own! This initial push contains the form, and the start of learning PHP to collect the form data...

Someday I'd like to make it an iOS app. I'd also like to make it more aesthetic, and make it so that others can customize a tracking system as well.